Tweet archive (2008 - 2015)
These are my 4,000+ tweets before deleting my account the first time. I
deleted it because it wasn't doing anything for me. At some point I
might go back and dig through these for what is worth keeping.

As far as twitter goes, I tried two more times to get back on but found
it to be more trouble than it was worth. It's nice to have an inbox that
is no longer pestering me with alerts about how stupid katy perry is.

:wq