Introduction Dynamic Components
`<components>` is a VUE reserved word. It is a wrapper to allow for
components to dynamically show. Use the `:is` attribute to assign which
component to use inside.

Components can also employ `<slots>`. In this commit, I use a slot for
`Quote` but not for `Author` or `New`. I also found it interesting to
realize that slots grab content by name and not by order. Good times.